+++ include/ignition/gazebo/components/AirPressureSensor.hh
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018 Open Source Robotics Foundation
+ * Copyright (C) 2019 Open Source Robotics Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,14 @@
  * limitations under the License.
  *
 */
-#ifndef IGNITION_GAZEBO_COMPONENTS_CAMERA_HH_
-#define IGNITION_GAZEBO_COMPONENTS_CAMERA_HH_
+#ifndef IGNITION_GAZEBO_COMPONENTS_AIRPRESSURE_HH_
+#define IGNITION_GAZEBO_COMPONENTS_AIRPRESSURE_HH_
 
 #include <sdf/Sensor.hh>
 
 #include <ignition/gazebo/components/Factory.hh>
 #include <ignition/gazebo/components/Component.hh>
+#include <ignition/gazebo/components/Serialization.hh>
 #include <ignition/gazebo/config.hh>
 
 namespace ignition
@@ -31,10 +32,11 @@ namespace gazebo
 inline namespace IGNITION_GAZEBO_VERSION_NAMESPACE {
 namespace components
 {
-  /// \brief A component type that contains a camera sensor,
-  /// sdf::Camera, information.
-  using Camera = Component<sdf::Sensor, class CameraTag>;
-  IGN_GAZEBO_REGISTER_COMPONENT("ign_gazebo_components.Camera", Camera)
+  /// \brief A component type that contains an air pressure sensor,
+  /// sdf::AirPressure, information.
+  using AirPressureSensor = Component<sdf::Sensor, class AirPressureSensorTag>;
+  IGN_GAZEBO_REGISTER_COMPONENT("ign_gazebo_components.AirPressureSensor",
+      AirPressureSensor)
 }
 }
 }