+++ src/systems/imu/Imu.cc
@@ -95,7 +95,8 @@ void Imu::PostUpdate(const UpdateInfo &_info,
     {
       // Update measurement time
       auto time = math::durationToSecNsec(_info.simTime);
-      it.second->Update(common::Time(time.first, time.second));
+      dynamic_cast<sensors::Sensor *>(it.second.get())->Update(
+          common::Time(time.first, time.second), false);
     }
   }
 
@@ -130,13 +131,13 @@ void ImuPrivate::CreateImuEntities(EntityComponentManager &_ecm)
       {
         // create sensor
         std::string sensorScopedName = scopedName(_entity, _ecm, "::", false);
-        auto data = _imu->Data()->Clone();
-        data->GetAttribute("name")->Set(sensorScopedName);
+        sdf::Sensor data = _imu->Data();
+        data.SetName(sensorScopedName);
         // check topic
-        if (!data->HasElement("topic"))
+        if (data.Topic().empty())
         {
           std::string topic = scopedName(_entity, _ecm) + "/imu";
-          data->GetElement("topic")->Set(topic);
+          data.SetTopic(topic);
         }
         std::unique_ptr<sensors::ImuSensor> sensor =
             this->sensorFactory.CreateSensor<